requirements.txt
django==2.0.5
django-summernote==0.8.8.7
django-modeltranslation==0.13-beta1
Pillow==5.1.0